openaibackend/requirements.txt
openai-agents
rich
pydantic
fastapi
uvicorn[standard]
python-dotenv
